+++ src/gitlab/merge_requests.c
@@ -72,7 +72,7 @@ gitlab_get_mrs(
         free(json_buffer.data);
         free(url);
         json_close(&stream);
-    } while ((url = next_url) && (max == -1 || count < max));
+    } while ((url = next_url) && (max == -1 || (int)count < max));
 
     free(url);
     free(e_owner);
